Relicense to AGPL
This is a simple project. Still, I don't want someone to rip it off and
host a copy with a bunch of ads or something unless they share the
source code of their modifications.

The AGPL makes sharing a requirement in such a case, as it considers
network use to be distribution.